creacion dto

+++ src/main/java/com/restaurante/menu_service/repository/DishRepository.java
@@ -3,5 +3,5 @@ package com.restaurante.menu_service.repository;
 import com.menu.service.entity.Dish;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DishRepository extends JpaRepository<Dish, Integer> {
+public interface DishRepository extends JpaRepository<Dish, Long> {
 }
